requirements.txt
flask
Flask-PyMongo
google-generativeai # Adjust the version according to the latest available
